Collect Photoshop IRB data as raw bytes while parsing
I have a need for the data in my project, and making a fork to
jpeg-decoder seemed like the easiest way to get access to it. There's a
PR in progress that would expose all App data from the JPEG as bytes,
which would make this hack a lot less useful.